Extract an abstract class for hash joins
