+++ enterprise/cypher/slotted-runtime/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_4/runtime/slotted/pipes/AbstractHashJoinPipe.scala
@@ -1,0 +1,81 @@
+/*
+ * Copyright (c) 2002-2017 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.cypher.internal.compatibility.v3_4.runtime.slotted.pipes
+
+import org.neo4j.cypher.internal.compatibility.v3_4.runtime.SlotConfiguration
+import org.neo4j.cypher.internal.compatibility.v3_4.runtime.slotted.PrimitiveExecutionContext
+import org.neo4j.cypher.internal.runtime.interpreted.ExecutionContext
+import org.neo4j.cypher.internal.runtime.interpreted.pipes.{Pipe, PipeWithSource, QueryState}
+
+import scala.collection.mutable
+
+abstract class AbstractHashJoinPipe[Key, T](left: Pipe,
+                                            right: Pipe,
+                                            slots: SlotConfiguration) extends PipeWithSource(left) {
+  val leftSide: T
+  val rightSide: T
+
+  override protected def internalCreateResults(input: Iterator[ExecutionContext], state: QueryState): Iterator[ExecutionContext] = {
+
+    if (input.isEmpty)
+      return Iterator.empty
+
+    val rhsIterator = right.createResults(state)
+
+    if (rhsIterator.isEmpty)
+      return Iterator.empty
+
+    val table = buildProbeTable(input)
+
+    if (table.isEmpty)
+      return Iterator.empty
+
+    val result = for {rhs: ExecutionContext <- rhsIterator
+                      joinKey <- computeKey(rhs, rightSide)}
+      yield {
+        val matchesFromLhs: mutable.Seq[ExecutionContext] = table.getOrElse(joinKey, mutable.MutableList.empty)
+
+        matchesFromLhs.map { lhs =>
+          val newRow = PrimitiveExecutionContext(slots)
+          lhs.copyTo(newRow)
+          copyDataFromRhs(newRow, rhs)
+          newRow
+        }
+      }
+
+    result.flatten
+  }
+
+  private def buildProbeTable(input: Iterator[ExecutionContext]): mutable.HashMap[Key, mutable.MutableList[ExecutionContext]] = {
+    val table = new mutable.HashMap[Key, mutable.MutableList[ExecutionContext]]
+
+    for {context <- input
+         joinKey <- computeKey(context, leftSide)} {
+      val matchingRows = table.getOrElseUpdate(joinKey, mutable.MutableList.empty)
+      matchingRows += context
+    }
+
+    table
+  }
+
+  def computeKey(context: ExecutionContext, keyColumns: T): Option[Key]
+
+  def copyDataFromRhs(newRow: PrimitiveExecutionContext, rhs: ExecutionContext): Unit
+}
